requirements.txt
django==2.2.10
django-environ==0.4.5
djangorestframework==3.11.0
psycopg2-binary==2.8.4
requests==2.23.0